+++ admin/js/book-review-admin-meta-box.js
@@ -53,7 +53,7 @@ jQuery(document).ready(function($) {
     };
 
     if (!$.trim($("#book_review_isbn").val())) {
-      $(".error-details").text("Please enter an ISBN.").show();
+      $(".error-details").text(book_review_google_api.no_isbn).show();
     }
     else {
       $(".error-details").hide();
@@ -67,19 +67,22 @@ jQuery(document).ready(function($) {
         try {
           json = JSON.parse(response);
 
-          if (json.status === "success") {
+          if (json.hasOwnProperty("data") && json.hasOwnProperty("status") && (json.status === "success")) {
             data = JSON.parse(json.data);
 
             // Check that data has been returned.
-            if (data.totalItems > 0 && data.items) {
+            if (data.hasOwnProperty("items") && (data.items.constructor === Array)
+              && (data.items.length > 0) && data.items[0].hasOwnProperty("volumeInfo")) {
               volume = data.items[0].volumeInfo;
 
               // Title
-              setField($("#book_review_title"), volume.title);
+              if (volume.hasOwnProperty("title")) {
+                setField($("#book_review_title"), volume.title);
+              }
 
               // Author
-              if ((volume.authors !== null) &&
-                !$.trim($("#book_review_author").val())) {
+              if (volume.hasOwnProperty("authors") && (volume.authors.constructor === Array)
+                && !$.trim($("#book_review_author").val())) {
                 for (i = 0; i < volume.authors.length; i++) {
                   author += volume.authors[i];
 
@@ -92,8 +95,8 @@ jQuery(document).ready(function($) {
               }
 
               // Genre
-              if ((volume.categories !== null) &&
-                !$.trim($("#book_review_genre").val())) {
+              if (volume.hasOwnProperty("categories") && (volume.categories.constructor === Array)
+                && !$.trim($("#book_review_genre").val())) {
                 for (i = 0; i < volume.categories.length; i++) {
                   genre += volume.categories[i];
 
@@ -106,27 +109,29 @@ jQuery(document).ready(function($) {
               }
 
               // Publisher
-              setField($("#book_review_publisher"), volume.publisher);
+              if (volume.hasOwnProperty("publisher")) {
+                setField($("#book_review_publisher"), volume.publisher);
+              }
 
               // Release Date
-              showFormattedDate(volume.publishedDate, json.format);
+              if (volume.hasOwnProperty("publishedDate")) {
+                showFormattedDate(volume.publishedDate, json.format);
+              }
 
               // Pages
-              setField($("#book_review_pages"), volume.pageCount);
+              if (volume.hasOwnProperty("pageCount")) {
+                setField($("#book_review_pages"), volume.pageCount);
+              }
 
               // Cover URL
-              if ((volume.imageLinks !== null) &&
-                (volume.imageLinks.thumbnail !== null) &&
+              if (volume.hasOwnProperty("imageLinks") && volume.imageLinks.hasOwnProperty("thumbnail") &&
                 !$.trim($("#book_review_cover_url").val())) {
-                if (volume.imageLinks.thumbnail) {
-                  $("#book_review_cover_url").val(volume.imageLinks.thumbnail);
-                  $("#book_review_cover_image").attr("src",
-                    volume.imageLinks.thumbnail).show();
-                }
+                $("#book_review_cover_url").val(volume.imageLinks.thumbnail);
+                $("#book_review_cover_image").attr("src", volume.imageLinks.thumbnail).show();
               }
 
               // Synopsis
-              if (volume.description !== null) {
+              if (volume.hasOwnProperty("description")) {
                 if (typeof tinymce !== "undefined") {
                   var editor = tinymce.get("book_review_summary");
 
@@ -145,28 +150,20 @@ jQuery(document).ready(function($) {
               }
             }
             else {
-              $(".error-details").text("A book with this ISBN was not found " +
-                "in the Google Books database.").show();
+              $(".error-details").text(book_review_google_api.not_found).show();
+              console.log(data);
             }
           }
-          else if (json.status === "error") {
-            $(".error-details").html("<p>Sorry, but something went wrong. Please check to ensure " +
-              "that you have created a Google API Key and that it has been entered correctly on " +
-              "the <em>Advanced</em> tab of the <em>Book Review Settings</em>.</p><p>Please also check " +
-              "to ensure that the correct IP address for your server has been entered into the " +
-              "<a href='https://code.google.com/apis/console' target='_blank'>Google Developers " +
-              "Console</a>. If in doubt, you may leave the IP address field empty. See the " +
-              "<a href='http://wpreviewplugins.com/book-review/#advanced' target='_blank'> " +
-              "documentation</a> for more information.</p><p>If you are still having trouble, " +
-              "please leave a message in the " +
-              "<a href='http://wpreviewplugins.com/support/forum/general-support/' " +
-              "target='_blank'>General Support forum</a>. Be sure to include the URL of your web " +
-              "site in your post. Thanks!").show();
-            console.log("Error message: " + json.data);
+          else if (json.hasOwnProperty("status") && (json.status === "error")) {
+            $(".error-details").html(book_review_google_api.unknown_error).show();
+
+            if (json.hasOwnProperty("data")) {
+              console.log("Error message: " + json.data);
+            }
           }
         }
         catch (e) {
-          console.log("JSON parsing error: ", e);
+          console.log("JSON parsing error: " + e);
         }
       });
     }
@@ -176,7 +173,7 @@ jQuery(document).ready(function($) {
 
   // Populate field if it does not already contain a value.
   function setField($elem, value) {
-    if (!$.trim($elem.val()) && (value !== null)) {
+    if (!$.trim($elem.val()) && (value != null)) {
       $elem.val(value);
     }
   }
@@ -184,8 +181,7 @@ jQuery(document).ready(function($) {
   function showFormattedDate(releaseDate, format) {
     var parts, currentDate, currentMonth, currentYear, months;
 
-    if (!$.trim($("#book_review_release_date").val()) &&
-      (releaseDate !== null)) {
+    if (!$.trim($("#book_review_release_date").val()) && (releaseDate != null)) {
       if (format !== "none") {
         parts = releaseDate.split("-");