+++ public/class-book-review-public.php
@@ -75,7 +75,7 @@ class Book_Review_Public {
    *
    *        Filters: http://codex.wordpress.org/Plugin_API#Filters
    *        Reference:  http://codex.wordpress.org/Plugin_API/Filter_Reference
-  *
+   *
    * @since    1.0.0
    *
    * @param    string    $content    Content of the post.